Move view tests to view test
